feat(#103 toast): device, user에 토스트 추가

+++ apps/v1/src/admin/pages/Asset/Management.tsx
@@ -19,6 +19,9 @@ export default function AssetPage() {
     const addToast = useToastStore((state) => state.addToast);
 
     const handleDelete = async (assetId: number) => {
+        const isConfirmed = window.confirm("선택한 항목을 삭제하시겠습니까?");
+        if (!isConfirmed) return;
+
         try {
             await deleteAsset(assetId);
             await mutate();
@@ -44,8 +47,7 @@ export default function AssetPage() {
     const AssetData = useAsset(data || [], handleDelete, handleEdit);
 
     const handleDeleteSelected = async () => {
-        const isConfirmed = window.confirm('정말 삭제하시겠습니까?');
-
+        const isConfirmed = window.confirm('선택한 항목을 삭제하시겠습니까?');
         if (!isConfirmed) return;
 
         if (selectedAssets.size === 0) {
